Changes for Theory Lessons 2.1
- Updating architectures to 64-bit
- Minimum deployment to iOS 7.1
- Fixing compiler warnings
- Adding "force pixel size" to SwiffView for iPhone 6 Plus

+++ Source/SwiffHTMLToCoreTextConverter.m
@@ -392,7 +392,7 @@ static void sParseNode(SwiffHTMLToCoreTextConverter *self, xmlNodePtr node)
     UInt8 *buffer = (UInt8 *)malloc(length);
     CFStringGetBytes(cfString, range, kCFStringEncodingUTF8, 0, 0, buffer, length, &length);
     
-    htmlDocPtr htmlDoc = htmlReadMemory((const char *)buffer, length, NULL, "UTF-8", 0);
+    htmlDocPtr htmlDoc = htmlReadMemory((const char *)buffer, (int)length, NULL, "UTF-8", 0);
    
     _characters     = [[NSMutableString alloc] init];
     _output         = [[NSMutableAttributedString alloc] init];
